@@ -4,16 +4,18 @@ import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
 import android.view.MenuItem
 import androidx.core.content.ContextCompat
+import com.connectycube.messenger.databinding.ActivityAttachmentPreviewBinding
 import com.connectycube.messenger.utilities.loadAttachImagePreview
-import kotlinx.android.synthetic.main.activity_attachment_preview.*
 
 const val EXTRA_URL = "attach_url"
 
 class AttachmentPreviewActivity : BaseChatActivity() {
+    private lateinit var binding: ActivityAttachmentPreviewBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_attachment_preview)
+        binding = ActivityAttachmentPreviewBinding.inflate(layoutInflater)
+        setContentView(binding.root)
         initToolBar()
         loadAttachment()
     }
@@ -34,10 +36,10 @@ class AttachmentPreviewActivity : BaseChatActivity() {
         val url = intent.getStringExtra(EXTRA_URL)
 
         if (url.isNullOrEmpty()) {
-            image_view.setImageResource(R.drawable.ic_image_black_24dp)
+            binding.imageView.setImageResource(R.drawable.ic_image_black_24dp)
             return
         }
-        loadAttachImagePreview(url, image_view, this)
+        loadAttachImagePreview(url, binding.imageView, this)
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
