@@ -7,7 +7,6 @@ import androidx.lifecycle.observe
 import com.connectycube.messenger.utilities.InjectorUtils
 import com.connectycube.messenger.viewmodels.SelectFromExistUsersViewModel
 import com.connectycube.messenger.vo.Status
-import kotlinx.android.synthetic.main.activity_create_chat.*
 
 const val EXTRA_USERS_TO_LOAD = "users_to_load"
 
@@ -26,10 +25,10 @@ open class SelectUsersFromExistActivity : SelectUsersBaseActivity<SelectFromExis
         getViewMode().getUsers(intent.extras!!.getIntegerArrayList(EXTRA_USERS_TO_LOAD)!!)
             .observe(this) { result ->
                 when (result.status) {
-                    Status.LOADING -> showProgress(progressbar)
-                    Status.ERROR -> hideProgress(progressbar)
+                    Status.LOADING -> showProgress(binding.progressbar)
+                    Status.ERROR -> hideProgress(binding.progressbar)
                     Status.SUCCESS -> {
-                        hideProgress(progressbar)
+                        hideProgress(binding.progressbar)
                         val users = result.data
                         setUsers(users)
                     }
